VP-1527: Move TrimStorePath to PathStringExtensions
