+++ VirtoCommerce.Storefront/Extensions/PathStringExtensions.cs
@@ -1,19 +1,16 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using VirtoCommerce.Storefront.Model;
 using VirtoCommerce.Storefront.Model.Stores;
 
 namespace VirtoCommerce.Storefront.Extensions
 {
-    public static class PathStringExtensions 
+    public static class PathStringExtensions
     {
         private static Regex _storeLangeExpr = new Regex(@"^/\b\S+\b/[a-zA-Z]{2}-[a-zA-Z]{2}/");
         public static PathString GetStoreAndLangSegment(this PathString path)
-        {       
+        {
             var matches = _storeLangeExpr.Match(path);
             return matches.Success ? matches.Value : "/";
         }
@@ -62,6 +59,37 @@ namespace VirtoCommerce.Storefront.Extensions
             result = result.Add(new PathString(path.TrimStoreAndLangSegment(store, language)));
 
             return result;
+        }
+
+        /// <summary>
+        /// Trims store path ("/store" for "http://localhost/store") from the beginning of the url. Does nothing in case of empty store.Url.
+        /// </summary>
+        /// <param name="path">Path to trim store path from.</param>
+        /// <param name="store">Store which path to trim.</param>
+        /// <returns></returns>
+        public static PathString TrimStorePath(this PathString path, Store store)
+        {
+            if (store == null)
+            {
+                throw new ArgumentNullException(nameof(store));
+            }
+
+            // Need to remove store path only if store has url
+            var storeUrl = !string.IsNullOrWhiteSpace(store.Url) ? store.Url : store.SecureUrl;
+
+            if (!string.IsNullOrWhiteSpace(storeUrl) && Uri.TryCreate(storeUrl, UriKind.Absolute, out var storeUri))
+            {
+                var storeUriPath = storeUri.AbsolutePath.Trim('/');
+
+                // Uri.AbsolutePath by default is "/" - no need to trim it
+                if (!string.IsNullOrWhiteSpace(storeUriPath) && !storeUriPath.Equals("/"))
+                {
+                    // Removing store url path from the beginning of path
+                    path = Regex.Replace(path, @"^/\b" + storeUriPath + @"\b/", "/", RegexOptions.IgnoreCase);
+                }
+            }
+
+            return path;
         }
     }
 }
